extract getMockName into a separate module
